Phase 7: Priority 2 Precision Fixes - Part 1 (2/5 files complete)
Fixed precision issues in:
1. ✅ test_forward_backward.c - Fixed test code bug (float* → double*)
2. ✅ cllm_symmetry.c - 17 instances fixed
3. ✅ cllm_root_word_modeling.c - 16 instances fixed

Changes:
- All float types converted to double
- All float literals (0.0f, 1.0f, etc.) converted to double (0.0, 1.0, etc.)
- Updated function signatures in headers
- Fixed test code memory corruption bug

Test Results:
- test_forward_backward: NOW PASSING (5/5 tests)
- Build: SUCCESS (0 errors, 11 warnings - no new warnings)
- All algorithm tests still passing

Progress:
- Priority 1: 100% complete (137 instances)
- Priority 2: 40% complete (2/5 files, ~33 instances fixed)
- Overall: ~170/303 instances fixed (56%)

Next: Continue with remaining Priority 2 files (cllm_positional.c, cllm_lattice_cache.c, cllm_lattice.c)

+++ include/cllm_attention.h
@@ -114,7 +114,7 @@ uint64_t cllm_get_token_prime(uint32_t token_id);
  * @param prime2 Second token's prime
  * @return Similarity score [0, 1]
  */
-float cllm_compute_prime_similarity(uint64_t prime1, uint64_t prime2);
+double cllm_compute_prime_similarity(uint64_t prime1, uint64_t prime2);
 
 /**
  * Apply symmetry operation to attention weights
@@ -175,8 +175,8 @@ void cllm_apply_fourier_dampening(float* attention_weights,
  * @param prime2 Second point's prime
  * @return Hyperdimensional distance
  */
-float cllm_compute_hyperdimensional_distance(const float* coords1,
-                                             const float* coords2,
+double cllm_compute_hyperdimensional_distance(const double* coords1,
+                                             const double* coords2,
                                              uint64_t prime1,
                                              uint64_t prime2);
