+++ src/ai/cllm_symmetry.c
@@ -19,16 +19,16 @@
  * @param angle Rotation angle in radians
  * @param dim Embedding dimension
  */
-static void apply_rotation(float* embedding, float angle, int dim) {
+static void apply_rotation(double* embedding, double angle, int dim) {
     if (!embedding || dim < 2) return;
     
-    float cos_a = prime_cos(angle);
-    float sin_a = prime_sin(angle);
+    double cos_a = prime_cos(angle);
+    double sin_a = prime_sin(angle);
     
     // Apply rotation in pairs of dimensions
     for (int i = 0; i < dim - 1; i += 2) {
-        float x = embedding[i];
-        float y = embedding[i + 1];
+        double x = embedding[i];
+        double y = embedding[i + 1];
         
         embedding[i] = cos_a * x - sin_a * y;
         embedding[i + 1] = sin_a * x + cos_a * y;
@@ -42,7 +42,7 @@ static void apply_rotation(float* embedding, float angle, int dim) {
  * @param axis Reflection axis (0 = x, 1 = y, 2 = z, etc.)
  * @param dim Embedding dimension
  */
-static void apply_reflection(float* embedding, int axis, int dim) {
+static void apply_reflection(double* embedding, int axis, int dim) {
     if (!embedding || axis < 0 || axis >= dim) return;
     
     // Reflect along specified axis
@@ -56,7 +56,7 @@ static void apply_reflection(float* embedding, int axis, int dim) {
  * @param scale Scale factor
  * @param dim Embedding dimension
  */
-static void apply_scaling(float* embedding, float scale, int dim) {
+static void apply_scaling(double* embedding, double scale, int dim) {
     if (!embedding || dim <= 0) return;
     
     for (int i = 0; i < dim; i++) {
@@ -82,13 +82,13 @@ uint32_t cllm_compute_symmetry_group(uint64_t prime) {
  * @param symmetry_group Symmetry group (0 to SYMMETRY_ORDER-1)
  * @param dim Embedding dimension
  */
-void cllm_apply_symmetry_transform(float* embedding, int symmetry_group, int dim) {
+void cllm_apply_symmetry_transform(double* embedding, int symmetry_group, int dim) {
     if (!embedding || symmetry_group < 0 || symmetry_group >= SYMMETRY_ORDER || dim <= 0) {
         return;
     }
     
     // Each symmetry group applies a different transformation
-    float angle = 2.0f * PI * (float)symmetry_group / (float)SYMMETRY_ORDER;
+    double angle = 2.0 * PI * (double)symmetry_group / (double)SYMMETRY_ORDER;
     
     switch (symmetry_group) {
         case 0:
@@ -116,7 +116,7 @@ void cllm_apply_symmetry_transform(float* embedding, int symmetry_group, int dim
         case 9:
             // Rotation + scaling
             apply_rotation(embedding, angle, dim);
-            apply_scaling(embedding, 1.1f, dim);
+            apply_scaling(embedding, 1.1, dim);
             break;
             
         case 6:
@@ -139,13 +139,13 @@ void cllm_apply_symmetry_transform(float* embedding, int symmetry_group, int dim
  * @param symmetry_group Symmetry group (0 to SYMMETRY_ORDER-1)
  * @param dim Embedding dimension
  */
-void cllm_apply_inverse_symmetry_transform(float* embedding, int symmetry_group, int dim) {
+void cllm_apply_inverse_symmetry_transform(double* embedding, int symmetry_group, int dim) {
     if (!embedding || symmetry_group < 0 || symmetry_group >= SYMMETRY_ORDER || dim <= 0) {
         return;
     }
     
     // Apply inverse transformation (reverse order, negative angles)
-    float angle = -2.0f * PI * (float)symmetry_group / (float)SYMMETRY_ORDER;
+    double angle = -2.0 * PI * (double)symmetry_group / (double)SYMMETRY_ORDER;
     
     switch (symmetry_group) {
         case 0:
@@ -172,7 +172,7 @@ void cllm_apply_inverse_symmetry_transform(float* embedding, int symmetry_group,
         case 3:
         case 9:
             // Inverse: scaling, then inverse rotation
-            apply_scaling(embedding, 1.0f / 1.1f, dim);
+            apply_scaling(embedding, 1.0 / 1.1, dim);
             apply_rotation(embedding, angle, dim);
             break;
             
@@ -198,13 +198,13 @@ void cllm_apply_inverse_symmetry_transform(float* embedding, int symmetry_group,
  * @param features Output features [num_features]
  * @param num_features Number of features to extract
  */
-void cllm_compute_symmetry_invariants(float* embedding, int dim, 
-                                     float* features, int num_features) {
+void cllm_compute_symmetry_invariants(double* embedding, int dim, 
+                                     double* features, int num_features) {
     if (!embedding || !features || dim <= 0 || num_features <= 0) return;
     
     // Feature 0: L2 norm (rotation invariant)
     if (num_features > 0) {
-        float norm = 0.0f;
+        double norm = 0.0;
         for (int i = 0; i < dim; i++) {
             norm += embedding[i] * embedding[i];
         }
@@ -213,7 +213,7 @@ void cllm_compute_symmetry_invariants(float* embedding, int dim,
     
     // Feature 1: Sum of absolute values (reflection invariant)
     if (num_features > 1) {
-        float sum = 0.0f;
+        double sum = 0.0;
         for (int i = 0; i < dim; i++) {
             sum += (embedding[i] >= 0) ? embedding[i] : -embedding[i];
         }
@@ -226,14 +226,14 @@ void cllm_compute_symmetry_invariants(float* embedding, int dim,
         for (int i = 0; i < dim; i++) {
             if (embedding[i] < 0) sign_product *= -1;
         }
-        features[2] = (float)sign_product;
+        features[2] = (double)sign_product;
     }
     
     // Feature 3: Maximum absolute value
     if (num_features > 3) {
-        float max_abs = 0.0f;
+        double max_abs = 0.0;
         for (int i = 0; i < dim; i++) {
-            float abs_val = (embedding[i] >= 0) ? embedding[i] : -embedding[i];
+            double abs_val = (embedding[i] >= 0) ? embedding[i] : -embedding[i];
             if (abs_val > max_abs) max_abs = abs_val;
         }
         features[3] = max_abs;
@@ -241,11 +241,11 @@ void cllm_compute_symmetry_invariants(float* embedding, int dim,
     
     // Additional features: moments
     for (int f = 4; f < num_features && f < 8; f++) {
-        float moment = 0.0f;
+        double moment = 0.0;
         int power = f - 2;
         for (int i = 0; i < dim; i++) {
-            float val = embedding[i];
-            float powered = val;
+            double val = embedding[i];
+            double powered = val;
             for (int p = 1; p < power; p++) {
                 powered *= val;
             }
@@ -264,26 +264,26 @@ void cllm_compute_symmetry_invariants(float* embedding, int dim,
  * @param transform_matrix Transformation matrix [dim x dim]
  * @param dim Embedding dimension
  */
-void cllm_apply_equivariant_transform(float* embedding, int symmetry_group,
-                                     float* transform_matrix, int dim) {
+void cllm_apply_equivariant_transform(double* embedding, int symmetry_group,
+                                     double* transform_matrix, int dim) {
     if (!embedding || !transform_matrix || dim <= 0) return;
     
     // First apply symmetry transformation
     cllm_apply_symmetry_transform(embedding, symmetry_group, dim);
     
     // Then apply linear transformation
-    float* temp = (float*)malloc(dim * sizeof(float));
+    double* temp = (double*)malloc(dim * sizeof(double));
     if (!temp) return;
     
     for (int i = 0; i < dim; i++) {
-        float sum = 0.0f;
+        double sum = 0.0;
         for (int j = 0; j < dim; j++) {
             sum += transform_matrix[i * dim + j] * embedding[j];
         }
         temp[i] = sum;
     }
     
-    memcpy(embedding, temp, dim * sizeof(float));
+    memcpy(embedding, temp, dim * sizeof(double));
     free(temp);
 }
 
@@ -298,31 +298,31 @@ void cllm_apply_equivariant_transform(float* embedding, int symmetry_group,
  * @param dim Embedding dimension
  * @return Compatibility score (0 to 1)
  */
-float cllm_symmetry_compatibility(float* embedding1, int symmetry1,
-                                 float* embedding2, int symmetry2, int dim) {
-    if (!embedding1 || !embedding2 || dim <= 0) return 0.0f;
+double cllm_symmetry_compatibility(double* embedding1, int symmetry1,
+                                 double* embedding2, int symmetry2, int dim) {
+    if (!embedding1 || !embedding2 || dim <= 0) return 0.0;
     
     // Create transformed copies
-    float* e1_transformed = (float*)malloc(dim * sizeof(float));
-    float* e2_transformed = (float*)malloc(dim * sizeof(float));
+    double* e1_transformed = (double*)malloc(dim * sizeof(double));
+    double* e2_transformed = (double*)malloc(dim * sizeof(double));
     
     if (!e1_transformed || !e2_transformed) {
         if (e1_transformed) free(e1_transformed);
         if (e2_transformed) free(e2_transformed);
-        return 0.0f;
+        return 0.0;
     }
     
-    memcpy(e1_transformed, embedding1, dim * sizeof(float));
-    memcpy(e2_transformed, embedding2, dim * sizeof(float));
+    memcpy(e1_transformed, embedding1, dim * sizeof(double));
+    memcpy(e2_transformed, embedding2, dim * sizeof(double));
     
     // Apply symmetry transformations
     cllm_apply_symmetry_transform(e1_transformed, symmetry1, dim);
     cllm_apply_symmetry_transform(e2_transformed, symmetry2, dim);
     
     // Compute cosine similarity
-    float dot = 0.0f;
-    float norm1 = 0.0f;
-    float norm2 = 0.0f;
+    double dot = 0.0;
+    double norm1 = 0.0;
+    double norm2 = 0.0;
     
     for (int i = 0; i < dim; i++) {
         dot += e1_transformed[i] * e2_transformed[i];
@@ -336,12 +336,12 @@ float cllm_symmetry_compatibility(float* embedding1, int symmetry1,
     norm1 = prime_sqrt(norm1);
     norm2 = prime_sqrt(norm2);
     
-    if (norm1 < 1e-8f || norm2 < 1e-8f) return 0.0f;
+    if (norm1 < 1e-8 || norm2 < 1e-8) return 0.0;
     
-    float similarity = dot / (norm1 * norm2);
+    double similarity = dot / (norm1 * norm2);
     
     // Map to [0, 1]
-    return (similarity + 1.0f) / 2.0f;
+    return (similarity + 1.0) / 2.0;
 }
 
 /**
@@ -352,14 +352,14 @@ float cllm_symmetry_compatibility(float* embedding1, int symmetry1,
  * @param seq_len Sequence length
  * @param mask Output attention mask [seq_len x seq_len]
  */
-void cllm_generate_symmetry_attention_mask(int* symmetry_groups, int seq_len, float* mask) {
+void cllm_generate_symmetry_attention_mask(int* symmetry_groups, int seq_len, double* mask) {
     if (!symmetry_groups || !mask || seq_len <= 0) return;
     
     for (int i = 0; i < seq_len; i++) {
         for (int j = 0; j < seq_len; j++) {
             // Tokens in same symmetry group have higher attention weight
             if (symmetry_groups[i] == symmetry_groups[j]) {
-                mask[i * seq_len + j] = 1.0f;
+                mask[i * seq_len + j] = 1.0;
             } else {
                 // Compute group distance
                 int dist = symmetry_groups[i] - symmetry_groups[j];
@@ -369,7 +369,7 @@ void cllm_generate_symmetry_attention_mask(int* symmetry_groups, int seq_len, fl
                 }
                 
                 // Decay based on distance
-                mask[i * seq_len + j] = 1.0f / (1.0f + 0.5f * (float)dist);
+                mask[i * seq_len + j] = 1.0 / (1.0 + 0.5 * (double)dist);
             }
         }
     }
